Fixed a generator bug, was missing a directory
If you were running the tinymce_installation generator for the first
time (not running to upgrade installation) and exception is raised because
files are trying to be copied into a directory that doesn't exist.

I've added the missing directory rule for the inlinepopups plugin to resolve
the issue.

Thanks to bking39116 for bringing this to my attention.