Separated url_path from install_path to support Rails apps in subdirectories.
The install_path (path from Rails_ROOT/public) won't match the url_path
to the TinyMCE installation when Rails is hosted in a subdirectory.
Tinymce::Hammer.url_path can now be set to resolve this issue.

This patch is based heavily on the excellent suggestion and patch submitted
by bbeardsley.