A ruby Hash is now accepted to Tinymce::Hammer.init
Previously only a 2D array of key value pairs was accepted.  This is
because ruby hashes are not ordered and certain init keys had to
appear before others.

If you use a hash your configuration values will be added in alphabetical
order by their stringified-keys.  If this does not work, you may still use
the array of arrays approach.